requirements.txt

asgiref
Babel

beautifulsoup4
brotli
calmjs.parse

cryptography
csscompressor

deepl

diff-match-patch

Django
django-admin-autocomplete-filter
django-allauth

django-clone
django-colorfield

django-db-connection-pool[postgresql]

django-import-export
django-phonenumber-field
django-postgrespool2
django-ratelimit
django-recaptcha

django-redis

django-redis-cache
django-select2
django-silk
django-simple-history
django-tinymce
django4-background-tasks

django_allauth

django_colorfield

django_compressor

django_cors-headers
django_debug_toolbar
django_imagekit
django_paypal

django_phonenumber_field
django_registration
django_request_profiler
django_safedelete
django_tinymce
gunicorn
hiredis

holidays
html5lib

humanize
inflection

jellyfish
lxml

olefile

pandas
pdfkit

phonenumbers
pilkit
pillow
playwright

polib
psycopg2-binary
pycountry
pyjwt

pytest
pytest-asyncio
pytest-dependency
pytest-django
pytest-playwright

pytest-xdist
python-dateutil

python-magic
python-redsys
python-slugify

python-telegram-bot
python3-openid
python3-saml
pytz

pyyaml
requests
requests-oauthlib

ruff

satispaython
setuptools
slugify
stripe

tzdata

whoosh
xhtml2pdf
